Saperated navbar in different component and added it in layout file.
