+++ app/components/Navbar/Navbar.js
@@ -1,0 +1,93 @@
+"use client";
+import { Button } from "@/components/ui/button";
+import { signOut } from "next-auth/react";
+import Link from "next/link";
+import React from "react";
+
+function Navbar() {
+  return (
+    <header className="flex h-16 w-full items-center justify-between bg-background px-4 md:px-6">
+      <Link
+        href="#"
+        className="flex items-center justify-center"
+        prefetch={false}
+      >
+        <MountainIcon className="h-6 w-6" />
+      </Link>
+      <h1 className="text-lg font-semi-bold">StreamLine</h1>
+      <div className="flex items-center gap-4">
+        {/* <Button variant="ghost" size="icon" className="lg:hidden">
+              <MenuIcon className="h-6 w-6" />
+              <span className="sr-only">Toggle navigation</span>
+            </Button> */}
+        <Button variant="outline" onClick={() => signOut()}>
+          <LogOutIcon className="h-5 w-5 md:mr-2 sm:mr-0" />
+          <div className="hidden sm:inline">Logout</div>
+        </Button>
+      </div>
+    </header>
+  );
+}
+
+export default Navbar;
+
+export function LogOutIcon(props) {
+  return (
+    <svg
+      {...props}
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4" />
+      <polyline points="16 17 21 12 16 7" />
+      <line x1="21" x2="9" y1="12" y2="12" />
+    </svg>
+  );
+}
+
+function MenuIcon(props) {
+  return (
+    <svg
+      {...props}
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <line x1="4" x2="20" y1="12" y2="12" />
+      <line x1="4" x2="20" y1="6" y2="6" />
+      <line x1="4" x2="20" y1="18" y2="18" />
+    </svg>
+  );
+}
+
+export function MountainIcon(props) {
+  return (
+    <svg
+      {...props}
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <path d="m8 3 4 8 5-5 5 15H2L8 3z" />
+    </svg>
+  );
+}
